pine64-pinephonepro: upgrade kernel to 5.16_rc4
This uses the Mobian kernel tree directly for now. We're working on
moving to a shared kernel in the pine64-org/linux Gitlab space again,
but till that has happened this kernel provides more hardware
functionality.

In my short testing this at least enables audio, the vibrator and the
battery indicator