main/postmarketos-hidden-desktop-entries: unhide htop, neovim, Glade, vim
I assume Glade was added here as it was always pulled in by some
dependency. This does not seem to be the case any more, so let's let
people who really want to use Glade on their phones do that
conveniently.

As for the terminal apps, I couldn't find any reason as to why they
were hidden in the first place and I'd rather like to have desktop
entries for them, so I'm proposing to unhide those too. None of them
should be preinstalled anyway I think.

Hiding of org.gnome.Extensions can probably be dropped when
https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/31086
has been merged.